# docs/BusAppSwagger.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/auth/profile":{"get":{"tags":["auth-rest-controller"],"operationId":"getUserProfile","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"put":{"tags":["auth-rest-controller"],"operationId":"updateUserProfile","parameters":[{"name":"newPassword","in":"query","required":false,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/bookings/submit":{"post":{"tags":["booking-rest-controller"],"operationId":"submitBooking","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookingDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/bookings/prepare-passenger-info":{"post":{"tags":["booking-rest-controller"],"operationId":"preparePassengerInfo","parameters":[{"name":"busScheduleId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"selectedSeats","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/auth/register":{"post":{"tags":["auth-rest-controller"],"operationId":"registerUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/admin/schedule":{"get":{"tags":["admin-bus-rest-controller"],"summary":"Get all Bus Schedules","operationId":"getAllSchedules","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BusScheduleDTO"}}}}}}},"post":{"tags":["admin-bus-rest-controller"],"summary":"Add a new Bus Schedule","operationId":"addSchedule","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BusScheduleDTO"}}},"required":true},"responses":{"201":{"description":"Schedule created successfully","content":{"*/*":{"schema":{"type":"object"}}}},"400":{"description":"Invalid schedule data"}}}},"/api/admin/bus":{"get":{"tags":["admin-bus-rest-controller"],"summary":"Get all Buses","operationId":"getAllBuses","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BusDTO"}}}}}}},"post":{"tags":["admin-bus-rest-controller"],"summary":"Add a new Bus","operationId":"addBus","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BusDTO"}}},"required":true},"responses":{"201":{"description":"Bus created successfully","content":{"*/*":{"schema":{"type":"string"}}}},"400":{"description":"Invalid bus data"}}}},"/api/buses/{id}":{"get":{"tags":["bus-rest-controller"],"operationId":"viewSchedule","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/buses/search":{"get":{"tags":["bus-rest-controller"],"operationId":"searchSchedules","parameters":[{"name":"source","in":"query","required":false,"schema":{"type":"string"}},{"name":"destination","in":"query","required":false,"schema":{"type":"string"}},{"name":"date","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/bookings/history":{"get":{"tags":["booking-rest-controller"],"operationId":"getBookingHistory","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Booking"}}}}}}}},"/api/bookings/form":{"get":{"tags":["booking-rest-controller"],"operationId":"getBookingForm","parameters":[{"name":"busScheduleId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/bookings/booked-seats":{"get":{"tags":["booking-rest-controller"],"operationId":"getBookedSeats","parameters":[{"name":"busScheduleId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"type":"string"}}}}}}}},"/api/auth/status":{"get":{"tags":["auth-rest-controller"],"operationId":"healthCheck","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/admin/dashboard":{"get":{"tags":["admin-bus-rest-controller"],"summary":"Get admin dashboard info","operationId":"showAdminDashboard","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"UserDTO":{"type":"object","properties":{"name":{"type":"string","minLength":1},"email":{"type":"string"},"password":{"type":"string","minLength":1}},"required":["name","password"]},"BookingDTO":{"type":"object","properties":{"busScheduleId":{"type":"integer","format":"int64"},"userEmail":{"type":"string"},"passengers":{"type":"array","items":{"$ref":"#/components/schemas/PassengerDTO"},"maxItems":2147483647,"minItems":1},"selectedSeats":{"type":"array","items":{"type":"string"}}},"required":["busScheduleId"]},"PassengerDTO":{"type":"object","properties":{"name":{"type":"string","minLength":1},"age":{"type":"integer","format":"int32","minimum":1},"gender":{"type":"string","minLength":1},"seatNumber":{"type":"string","minLength":1}},"required":["gender","name","seatNumber"]},"BusScheduleDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"busId":{"type":"integer","format":"int64"},"busName":{"type":"string"},"source":{"type":"string"},"destination":{"type":"string"},"date":{"type":"string","format":"date"},"departureTime":{"type":"string"},"arrivalTime":{"type":"string"},"price":{"type":"number","format":"double"},"availableSeats":{"type":"integer","format":"int32"},"totalSeats":{"type":"integer","format":"int32"}}},"BusDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"busName":{"type":"string"},"owner":{"type":"string"},"type":{"type":"string"},"totalSeats":{"type":"integer","format":"int32"},"schedules":{"type":"array","items":{"$ref":"#/components/schemas/BusScheduleDTO"}}}},"Booking":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"bookingTime":{"type":"string","format":"date-time"},"numberOfSeats":{"type":"integer","format":"int32"},"totalFare":{"type":"number","format":"double"},"user":{"$ref":"#/components/schemas/User"},"passengers":{"type":"array","items":{"$ref":"#/components/schemas/Passenger"}},"busSchedule":{"$ref":"#/components/schemas/BusSchedule"}}},"Bus":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"busName":{"type":"string"},"owner":{"type":"string"},"type":{"type":"string"},"totalSeats":{"type":"integer","format":"int32"},"schedules":{"type":"array","items":{"$ref":"#/components/schemas/BusSchedule"}}}},"BusSchedule":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"bus":{"$ref":"#/components/schemas/Bus"},"source":{"type":"string"},"destination":{"type":"string"},"date":{"type":"string","format":"date"},"departureTime":{"type":"string"},"arrivalTime":{"type":"string"},"price":{"type":"number","format":"double"},"totalSeats":{"type":"integer","format":"int32"},"bookedSeats":{"type":"array","items":{"type":"string"},"uniqueItems":true}}},"Passenger":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"age":{"type":"integer","format":"int32"},"gender":{"type":"string"},"seatNumber":{"type":"string"},"booking":{"$ref":"#/components/schemas/Booking"}}},"User":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"role":{"type":"string"}}}}}}